requirements.txt
numpy
openai
pytest
pytest-asyncio
python-dotenv